fixed undo bug
In games with multiple deal piles, where the number of cards
in the deck was not divisible by the number of piles, the final deal
would not spread out across all the piles but would "undo" by taking
a card from every pile, resulting in a shift in card placement.
This was true for pyramid pairs draw three.